Lexer partial rewrite
