+++ runtime/Rust/src/lex/lexer.rs
@@ -1,280 +1,80 @@
 //! Lexer implementation
-use std::borrow::Cow::Borrowed;
-use std::borrow::{Borrow, Cow};
-use std::cell::{Cell, RefCell};
 
-use std::rc::Rc;
+use std::{collections::VecDeque, io::{Cursor, Read}};
 
-use crate::char_stream::{CharStream, InputData};
-use crate::error_listener::{ConsoleErrorListener, ErrorListener};
-use crate::errors::ANTLRError;
-use crate::int_stream::IntStream;
-use crate::lexer_atn_simulator::{ILexerATNSimulator, LexerATNSimulator};
-use crate::parser::ParserNodeType;
+use crate::lex::{Token, action::LexerAction, lexer_atn_simulator::LexerATNSimulator, token::{TokenChannel, TokenType}};
 
-use crate::recognizer::{Actions, Recognizer};
-use crate::rule_context::EmptyContextType;
-use crate::token::TOKEN_INVALID_TYPE;
-use crate::token_factory::{CommonTokenFactory, TokenAware, TokenFactory};
-use crate::token_source::TokenSource;
-use std::ops::{Deref, DerefMut};
-
-///  Lexer functionality required by `LexerATNSimulator` to work properly
-pub trait Lexer<'input>:
-    TokenSource<'input>
-    + Recognizer<'input, Node = EmptyContextType<'input, <Self as TokenAware<'input>>::TF>>
-{
-    /// Concrete input stream used by this parser
-    type Input: IntStream;
-    /// Same as `TokenStream::get_input_stream` but returns concrete type instance
-    /// important for proper inlining in hot code of `LexerATNSimulator`
-    fn input(&mut self) -> &mut Self::Input;
-    /// Sets channel where current token will be pushed
-    ///
-    /// By default two channels are available:
-    ///  - `LEXER_DEFAULT_TOKEN_CHANNEL`
-    ///  - `LEXER_HIDDEN`
-    fn set_channel(&mut self, v: i32);
-
-    /// Pushes current mode to internal mode stack and sets `m` as current lexer mode
-    /// `pop_mode should be used to recover previous mode
-    fn push_mode(&mut self, m: usize);
-
-    /// Pops mode from internal mode stack
-    fn pop_mode(&mut self) -> Option<usize>;
-
-    /// Sets type of the current token
-    /// Called from action to override token that will be emitted by lexer
-    fn set_type(&mut self, t: i32);
-
-    /// Sets lexer mode discarding current one
-    fn set_mode(&mut self, m: usize);
-
-    /// Used to informs lexer that it should consider next token as a continuation of the current one
-    fn more(&mut self);
-
-    /// Tells lexer to completely ignore and not emit current token.
-    fn skip(&mut self);
-
-    #[doc(hidden)]
-    fn reset(&mut self);
-
-    #[doc(hidden)]
-    fn get_interpreter(&self) -> Option<&LexerATNSimulator>;
+#[derive(Debug)]
+struct LexerPosition {
+    pub absolute: usize,
+    pub line: usize,
+    pub char_position_in_line: usize,
 }
 
-/// **! Usually generated by ANTLR !**
-///
-/// This trait combines everything that can be used to extend Lexer behavior
-pub trait LexerRecog<'a, T: Recognizer<'a>>: Actions<'a, T> + Sized + 'static {
-    /// Callback to extend emit behavior
-    fn before_emit(_lexer: &mut T) {}
+pub(crate) enum LexerAction {
+    LexerChannelAction(i32),
+    LexerCustomAction {
+        rule_index: i32,
+        action_index: i32,
+    },
+    LexerModeAction(i32),
+    LexerMoreAction,
+    LexerPopModeAction,
+    LexerPushModeAction(i32),
+    LexerSkipAction,
+    LexerTypeAction(i32),
+    LexerIndexedCustomAction {
+        offset: isize,
+        action: Box<LexerAction>,
+    },
 }
 
 /// Default implementation of Lexer
 ///
 /// Public fields in this struct are intended to be used by embedded actions
 #[allow(missing_docs)]
-pub struct BaseLexer<
-    'input,
-    T: LexerRecog<'input, Self> + 'static,
-    Input: CharStream<TF::From>,
-    TF: TokenFactory<'input> = CommonTokenFactory,
-> {
-    /// `LexerATNSimulator` instance of this lexer
-    pub interpreter: Option<Box<LexerATNSimulator>>,
-    /// `CharStream` used by this lexer
-    pub input: Option<Input>,
-    recog: T,
+pub struct Lexer<'input> {
+    input: Cursor<&'input str>,
+    // position: LexerPosition,
+    interpreter: LexerATNSimulator,
+    
+    // error_listeners: Vec<Box<dyn ErrorListener<'input, Self>>>,
 
-    factory: &'input TF,
+    // current_pos: LexerPosition,
 
-    error_listeners: RefCell<Vec<Box<dyn ErrorListener<'input, Self>>>>,
-
-    pub token_start_char_index: isize,
-    pub token_start_line: isize,
-    pub token_start_column: isize,
-    current_pos: Rc<LexerPosition>,
     /// Overrides token type emitted by lexer for current token
-    pub token_type: i32,
-    /// Make it `Some` to override token that is currently being generated by lexer
-    pub token: Option<TF::Tok>,
+    token_type: TokenType,
+
     hit_eof: bool,
     /// Channel lexer is currently assigning tokens to
-    pub channel: i32,
+    channel: TokenChannel,
     /// stack of modes, which is used for pushMode,popMode lexer actions
-    pub mode_stack: Vec<usize>,
-    /// Mode lexer is currently in
-    pub mode: usize,
-    /// Make it `Some` to override text for token that is currently being generated by lexer
-    pub text: Option<<TF::Data as ToOwned>::Owned>,
+    mode_stack: VecDeque<LexerMode>,
+    mode: LexerMode,
+
+    /// Make it `Some` to override token that is currently being generated by lexer
+    force_next_token: Option<Token<'input>>,
 }
 
-#[derive(Debug)]
-pub(crate) struct LexerPosition {
-    pub(crate) line: Cell<isize>,
-    pub(crate) char_position_in_line: Cell<isize>,
+pub enum LexerMode {
+    Default,
+
+    /// Special token type to indicate that lexer should continue current token on next iteration
+    More,
+
+    /// Special token type to indicate that lexer should not return current token
+    /// usually used to skip whitespaces and comments
+    Skip
 }
-
-impl<'input, T, Input, TF> Deref for BaseLexer<'input, T, Input, TF>
-where
-    T: LexerRecog<'input, Self> + 'static,
-    Input: CharStream<TF::From>,
-    TF: TokenFactory<'input>,
-{
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        &self.recog
-    }
-}
-
-impl<'input, T, Input, TF> DerefMut for BaseLexer<'input, T, Input, TF>
-where
-    T: LexerRecog<'input, Self> + 'static,
-    Input: CharStream<TF::From>,
-    TF: TokenFactory<'input>,
-{
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.recog
-    }
-}
-
-impl<'input, T, Input, TF> Recognizer<'input> for BaseLexer<'input, T, Input, TF>
-where
-    T: LexerRecog<'input, Self> + 'static,
-    Input: CharStream<TF::From>,
-    TF: TokenFactory<'input>,
-{
-    type Node = EmptyContextType<'input, TF>;
-
-    fn sempred(
-        &mut self,
-        _localctx: Option<&<Self::Node as ParserNodeType<'input>>::Type>,
-        rule_index: i32,
-        action_index: i32,
-    ) -> bool {
-        <T as Actions<'input, Self>>::sempred(_localctx, rule_index, action_index, self)
-    }
-
-    fn action(
-        &mut self,
-        _localctx: Option<&<Self::Node as ParserNodeType<'input>>::Type>,
-        rule_index: i32,
-        action_index: i32,
-    ) {
-        <T as Actions<'input, Self>>::action(_localctx, rule_index, action_index, self)
-    }
-}
-
-/// Default lexer mode id
-pub const LEXER_DEFAULT_MODE: usize = 0;
-/// Special token type to indicate that lexer should continue current token on next iteration
-/// see `Lexer::more()`
-pub const LEXER_MORE: i32 = -2;
-/// Special token type to indicate that lexer should not return current token
-/// usually used to skip whitespaces and comments
-/// see `Lexer::skip()`
-pub const LEXER_SKIP: i32 = -3;
-
-#[doc(inline)]
-pub use super::token::TOKEN_DEFAULT_CHANNEL as LEXER_DEFAULT_TOKEN_CHANNEL;
-
-#[doc(inline)]
-pub use super::token::TOKEN_HIDDEN_CHANNEL as LEXER_HIDDEN;
 
 pub(crate) const LEXER_MIN_CHAR_VALUE: i32 = 0x0000;
 pub(crate) const LEXER_MAX_CHAR_VALUE: i32 = 0x10FFFF;
 
-impl<'input, T, Input, TF> BaseLexer<'input, T, Input, TF>
-where
-    T: LexerRecog<'input, Self> + 'static,
-    Input: CharStream<TF::From>,
-    TF: TokenFactory<'input>,
+impl<'input> Lexer<'input>
 {
-    fn emit_token(&mut self, token: TF::Tok) {
-        self.token = Some(token);
-    }
-
-    fn emit(&mut self) {
-        <T as LexerRecog<Self>>::before_emit(self);
-        let stop = self.get_char_index() - 1;
-        let token = self.factory.create(
-            Some(self.input.as_mut().unwrap()),
-            self.token_type,
-            self.text.take(),
-            self.channel,
-            self.token_start_char_index,
-            stop,
-            self.token_start_line,
-            self.token_start_column,
-        );
-        self.emit_token(token);
-    }
-
-    fn emit_eof(&mut self) {
-        let token = self.factory.create(
-            None::<&mut Input>,
-            super::int_stream::EOF,
-            None,
-            LEXER_DEFAULT_TOKEN_CHANNEL,
-            self.get_char_index(),
-            self.get_char_index() - 1,
-            self.get_line(),
-            self.get_char_position_in_line(),
-        );
-        self.emit_token(token)
-    }
-
-    /// Current position in input stream
-    pub fn get_char_index(&self) -> isize {
-        self.input.as_ref().unwrap().index()
-    }
-
-    /// Current token text
-    pub fn get_text<'a>(&'a self) -> Cow<'a, TF::Data>
-    where
-        'input: 'a,
-    {
-        self.text
-            .as_ref()
-            .map(|it| Borrowed(it.borrow()))
-            // .unwrap_or("")
-            .unwrap_or_else(|| {
-                let text = self
-                    .input
-                    .as_ref()
-                    .unwrap()
-                    .get_text(self.token_start_char_index, self.get_char_index() - 1);
-                TF::get_data(text)
-            })
-    }
-
-    /// Used from lexer actions to override text of the token that will be emitted next
-    pub fn set_text(&mut self, _text: <TF::Data as ToOwned>::Owned) {
-        self.text = Some(_text);
-    }
-
-    // fn get_all_tokens(&mut self) -> Vec<TF::Tok> { unimplemented!() }
-
-    // fn get_char_error_display(&self, _c: char) -> String { unimplemented!() }
-
-    /// Add error listener
-    pub fn add_error_listener(&mut self, listener: Box<dyn ErrorListener<'input, Self>>) {
-        self.error_listeners.borrow_mut().push(listener);
-    }
-
-    /// Remove and drop all error listeners
-    pub fn remove_error_listeners(&mut self) {
-        self.error_listeners.borrow_mut().clear();
-    }
-
     /// Creates new lexer instance
-    pub fn new_base_lexer(
-        input: Input,
-        interpreter: LexerATNSimulator,
-        recog: T,
-        factory: &'input TF,
+    pub fn new(
+        input: &'input str
     ) -> Self {
         let mut lexer = Self {
             interpreter: Some(Box::new(interpreter)),
@@ -291,9 +91,9 @@ where
             }),
             token_type: super::token::TOKEN_INVALID_TYPE,
             text: None,
-            token: None,
+            force_next_token: None,
             hit_eof: false,
-            channel: super::token::TOKEN_DEFAULT_CHANNEL,
+            channel: super::token::TokenChannel::Default,
             //            token_factory_source_pair: None,
             mode_stack: Vec::new(),
             mode: self::LEXER_DEFAULT_MODE,
@@ -302,163 +102,17 @@ where
         lexer.interpreter.as_mut().unwrap().current_pos = pos;
         lexer
     }
-}
+    
 
-impl<'input, T, Input, TF> TokenAware<'input> for BaseLexer<'input, T, Input, TF>
-where
-    T: LexerRecog<'input, Self> + 'static,
-    Input: CharStream<TF::From>,
-    TF: TokenFactory<'input>,
-{
-    type TF = TF;
-}
+    /// Add error listener
+    // pub fn add_error_listener(&mut self, listener: Box<dyn ErrorListener<'input, Self>>) {
+    //     self.error_listeners.borrow_mut().push(listener);
+    // }
 
-impl<'input, T, Input, TF> TokenSource<'input> for BaseLexer<'input, T, Input, TF>
-where
-    T: LexerRecog<'input, Self> + 'static,
-    Input: CharStream<TF::From>,
-    TF: TokenFactory<'input>,
-{
-    type TF = TF;
-    #[inline]
-    #[allow(unused_labels)]
-    fn next_token(&mut self) -> <Self::TF as TokenFactory<'input>>::Tok {
-        assert!(self.input.is_some());
-
-        let _marker = self.input().mark();
-        'outer: loop {
-            if self.hit_eof {
-                self.emit_eof();
-                break;
-            }
-            self.token = None;
-            self.channel = LEXER_DEFAULT_TOKEN_CHANNEL;
-            self.token_start_column = self
-                .interpreter
-                .as_ref()
-                .unwrap()
-                .get_char_position_in_line();
-            self.token_start_line = self.interpreter.as_ref().unwrap().get_line();
-            self.text = None;
-            let index = self.input().index();
-            self.token_start_char_index = index;
-
-            'inner: loop {
-                self.token_type = TOKEN_INVALID_TYPE;
-                // detach from self, to allow self to be passed deeper
-                let mut interpreter = self.interpreter.take().unwrap();
-                //                    let mut input = self.input.take().unwrap();
-                let result = interpreter.match_token(self.mode, self);
-                self.interpreter = Some(interpreter);
-
-                let ttype = result.unwrap_or_else(|err| {
-                    //                            println!("error, recovering");
-                    notify_listeners(&mut self.error_listeners.borrow_mut(), &err, self);
-                    self.interpreter
-                        .as_mut()
-                        .unwrap()
-                        .recover(err, self.input.as_mut().unwrap());
-                    LEXER_SKIP
-                });
-                //                    self.input = Some(input)
-
-                if self.input().la(1) == super::int_stream::EOF {
-                    self.hit_eof = true;
-                }
-
-                if self.token_type == TOKEN_INVALID_TYPE {
-                    self.token_type = ttype;
-                }
-
-                if self.token_type == LEXER_SKIP {
-                    continue 'outer;
-                }
-
-                if self.token_type != LEXER_MORE {
-                    break;
-                }
-            }
-
-            if self.token.is_none() {
-                self.emit();
-                break;
-            }
-        }
-        self.input().release(_marker);
-        self.token.take().unwrap()
+    /// Remove and drop all error listeners
+    pub fn remove_error_listeners(&mut self) {
+        self.error_listeners.borrow_mut().clear();
     }
-
-    fn get_line(&self) -> isize {
-        self.current_pos.line.get()
-    }
-
-    fn get_char_position_in_line(&self) -> isize {
-        self.current_pos.char_position_in_line.get()
-    }
-
-    fn get_input_stream(&mut self) -> Option<&mut dyn IntStream> {
-        self.input.as_mut().map(|x| x as _)
-    }
-
-    fn get_source_name(&self) -> String {
-        self.input
-            .as_ref()
-            .map(|it| it.get_source_name())
-            .unwrap_or("<none>".to_string())
-    }
-
-    //    fn set_token_factory<'c: 'b>(&mut self, f: &'c TokenFactory) {
-    //        self.factory = f;
-    //    }
-
-    fn get_token_factory(&self) -> &'input TF {
-        self.factory
-    }
-
-    fn get_dfa_string(&self) -> String {
-        self.get_interpreter().unwrap().get_dfa_for_mode(LEXER_DEFAULT_MODE).upgradable_read().to_lexer_string()
-    }
-}
-
-#[cold]
-#[inline(never)]
-fn notify_listeners<'input, T, Input, TF>(
-    liseners: &mut Vec<Box<dyn ErrorListener<'input, BaseLexer<'input, T, Input, TF>>>>,
-    e: &ANTLRError,
-    lexer: &BaseLexer<'input, T, Input, TF>,
-) where
-    T: LexerRecog<'input, BaseLexer<'input, T, Input, TF>> + 'static,
-    Input: CharStream<TF::From>,
-    TF: TokenFactory<'input>,
-{
-    let inner = lexer
-        .input
-        .as_ref()
-        .unwrap()
-        .get_text(lexer.token_start_char_index, lexer.get_char_index());
-    let text = format!(
-        "token recognition error at: '{}'",
-        TF::get_data(inner).to_display()
-    );
-    for listener in liseners.iter_mut() {
-        listener.syntax_error(
-            lexer,
-            None,
-            lexer.token_start_line,
-            lexer.token_start_column,
-            &text,
-            Some(e),
-        )
-    }
-}
-
-impl<'input, T, Input, TF> Lexer<'input> for BaseLexer<'input, T, Input, TF>
-where
-    T: LexerRecog<'input, Self> + 'static,
-    Input: CharStream<TF::From>,
-    TF: TokenFactory<'input>,
-{
-    type Input = Input;
 
     fn input(&mut self) -> &mut Self::Input {
         self.input.as_mut().unwrap()
@@ -488,11 +142,11 @@ where
     }
 
     fn more(&mut self) {
-        self.set_type(LEXER_MORE)
+        self.set_type(LexerMode::More)
     }
 
     fn skip(&mut self) {
-        self.set_type(LEXER_SKIP)
+        self.set_type(LexerMode::Skip)
     }
 
     fn reset(&mut self) {
@@ -502,4 +156,235 @@ where
     fn get_interpreter(&self) -> Option<&LexerATNSimulator> {
         self.interpreter.as_deref()
     }
+
+    #[inline]
+    fn consume(&mut self) {
+        self.base.consume();
+        //        self.base.p = self.next_token_on_channel(self.base.p,self.channel);
+        //        self.base.current_token_index = self.base.p;
+        let next = self.next_token_on_channel(self.base.p, self.channel, 1);
+        self.base.seek(next);
+        // Ok(())
+    }
+
+    #[inline]
+    fn la(&mut self, i: isize) -> i32 {
+        self.lt(i)
+            .map(|t| t.borrow().get_token_type())
+            .unwrap_or(TOKEN_INVALID_TYPE)
+    }
+
+    #[inline(always)]
+    fn mark(&mut self) -> isize {
+        0
+    }
+
+    #[inline(always)]
+    fn release(&mut self, _marker: isize) {}
+
+    #[inline(always)]
+    fn index(&self) -> isize {
+        self.base.index()
+    }
+
+    #[inline(always)]
+    fn seek(&mut self, index: isize) {
+        self.base.seek(index);
+    }
+
+    #[inline(always)]
+    fn size(&self) -> isize {
+        self.base.size()
+    }
+
+    fn get_source_name(&self) -> String {
+        self.base.get_source_name()
+    }
+
+    #[inline(always)]
+    fn lt(&mut self, k: isize) -> Option<Token> {
+        if k == 0 {
+            return None;
+        }
+        if k < 0 {
+            return self.lb(-k);
+        }
+        self.lt_inner(k)
+    }
+
+    #[inline]
+    fn get(&self, index: isize) -> &<Self::TF as TokenFactory<'input>>::Tok {
+        self.base.get(index)
+    }
+
+    fn get_token_source(&self) -> &dyn TokenSource<'input, TF = Self::TF> {
+        self.base.get_token_source()
+    }
+
+    fn get_text_from_interval(&self, start: isize, stop: isize) -> String {
+        self.base.get_text_from_interval(start, stop)
+    }
+    
+    pub fn get_dfa_string(&self) -> String {
+        self.base.get_dfa_string()
+    }
+
+    fn lt_inner(&mut self, k: isize) -> Option<Token> {
+        let mut i = self.base.p;
+        let mut n = 1; // we know tokens[p] is a good one
+                       // find k good tokens
+        while n < k {
+            // skip off-channel tokens, but make sure to not look past EOF
+            if self.sync(i + 1) {
+                i = self.next_token_on_channel(i + 1, self.channel, 1);
+            }
+            n += 1;
+        }
+        //		if ( i>range ) range = i;
+        self.base.tokens.get(i as usize)
+    }
+
+    /// Restarts this token stream
+    pub fn reset(&mut self) {
+        self.base.p = 0;
+        self.base.current_token_index = 0;
+    }
+
+    fn sync(&mut self, i: isize) -> bool {
+        let need = i - self.size() + 1;
+        if need > 0 {
+            let fetched = self.base.fill(need);
+            return fetched >= need;
+        }
+
+        true
+    }
+
+    //todo make this const generic over direction
+    fn next_token_on_channel(&mut self, mut i: isize, channel: i32, direction: isize) -> isize {
+        self.sync(i);
+        if i >= self.size() {
+            return self.size() - 1;
+        }
+
+        let mut token = self.base.tokens[i as usize].borrow();
+        while token.get_channel() != channel {
+            i += direction;
+            if token.get_token_type() == EOF || i < 0 {
+                return i;
+            }
+
+            self.sync(i);
+            token = self.base.tokens[i as usize].borrow();
+        }
+
+        i
+    }
+
+    /// Scans backwards from index `i` looking for a token on the given
+    /// `channel`. Returns the index of the first matching token, or -1 if
+    /// none is found before the start of the stream. EOF tokens are treated
+    /// as matching any channel.
+    fn previous_token_on_channel(&mut self, mut i: isize, channel: i32) -> isize {
+        self.sync(i);
+        if i >= self.size() {
+            return self.size() - 1;
+        }
+
+        while i >= 0 {
+            let token = self.base.tokens[i as usize].borrow();
+            if token.get_token_type() == EOF || token.get_channel() == channel {
+                return i;
+            }
+            i -= 1;
+        }
+
+        -1
+    }
+
+    /// Collects tokens in the index range `[from, to]` that are on the
+    /// given `channel`. If `channel` is -1, collects all tokens that are
+    /// *not* on `TokenChannel::Default`. Returns `None` if no tokens match.
+    fn filter_for_channel(&self, from: isize, to: isize, channel: i32) -> Option<Vec<OwningToken>> {
+        let mut hidden: Vec<OwningToken> = Vec::new();
+        for i in from..=to {
+            if i < 0 || i >= self.base.tokens.len() as isize {
+                continue;
+            }
+            let t = self.base.tokens[i as usize].borrow();
+            if channel == -1 {
+                if t.get_channel() != TokenChannel::Default {
+                    hidden.push(t.to_owned());
+                }
+            } else if t.get_channel() == channel {
+                hidden.push(t.to_owned());
+            }
+        }
+
+        if hidden.is_empty() {
+            None
+        } else {
+            Some(hidden)
+        }
+    }
+
+    /// Returns hidden tokens to the left of `token_index` on the given
+    /// `channel`. Scans backwards to the previous on-channel token and
+    /// collects everything in between. Pass -1 as `channel` to get all
+    /// non-default-channel tokens.
+    pub fn get_hidden_tokens_to_left(
+        &mut self,
+        token_index: isize,
+        channel: i32,
+    ) -> Option<Vec<OwningToken>> {
+        let prev_on_channel =
+            self.previous_token_on_channel(token_index - 1, TokenChannel::Default);
+        let from = if prev_on_channel == token_index - 1 || prev_on_channel < 0 {
+            return None;
+        } else {
+            prev_on_channel + 1
+        };
+        self.filter_for_channel(from, token_index - 1, channel)
+    }
+
+    /// Returns hidden tokens to the right of `token_index` on the given
+    /// `channel`. Scans forward to the next on-channel token and collects
+    /// everything in between. Pass -1 as `channel` to get all
+    /// non-default-channel tokens.
+    pub fn get_hidden_tokens_to_right(
+        &mut self,
+        token_index: isize,
+        channel: i32,
+    ) -> Option<Vec<Token>> {
+        let next_on_channel = self.next_token_on_channel(token_index + 1, TokenChannel::Default, 1);
+        let to = next_on_channel - 1;
+        if to <= token_index {
+            return None;
+        }
+        self.filter_for_channel(token_index + 1, to, channel)
+    }
+
+    // Stands for look-back?
+    fn lb(
+        &mut self,
+        k: isize,
+    ) -> Option<Token> {
+        if k == 0 || (self.base.p - k) < 0 {
+            return None;
+        }
+
+        let mut i = self.base.p;
+        let mut n = 1;
+        // find k good tokens looking backwards
+        while n <= k && i > 0 {
+            // skip off-channel tokens
+            i = self.next_token_on_channel(i - 1, self.channel, -1);
+            n += 1;
+        }
+        if i < 0 {
+            return None;
+        }
+
+        self.base.tokens.get(i as usize)
+    }    
 }
