+++ runtime/Rust/src/lib.rs
@@ -71,181 +71,7 @@
 //! [`TokenStream`]: crate::token_stream::TokenStream
 //! [`ParserRuleContext`]: crate::parser_rule_context::ParserRuleContext
 
-#[macro_use]
-extern crate lazy_static;
-
-#[doc(hidden)]
-pub use lazy_static::lazy_static;
-
-#[doc(hidden)]
-pub use parking_lot::RwLock;
-
-#[doc(hidden)]
-pub use better_any::{tid, Tid, TidAble, TidExt};
-
-#[doc(inline)]
-pub use error_strategy::{BailErrorStrategy, DefaultErrorStrategy, ErrorStrategy};
-
-pub use input_stream::InputStream;
-
-#[doc(inline)]
-pub use lexer::{BaseLexer, Lexer};
-#[doc(inline)]
-pub use parser::{BaseParser, ListenerId, Parser};
-#[doc(inline)]
-pub use token_source::TokenSource;
-//extern crate uuid;
-#[doc(hidden)]
-pub use prediction_context::PredictionContextCache;
-
-#[doc(inline)]
-pub use prediction_mode::PredictionMode;
-
-#[doc(hidden)]
-pub mod atn_config;
-#[doc(hidden)]
-pub mod atn_simulator;
-pub mod int_stream;
-mod lexer_action;
-mod ll1_analyzer;
-#[doc(hidden)]
-pub mod recognizer;
-pub mod token_factory;
-#[doc(hidden)]
-#[doc(hidden)]
-pub mod atn_state;
-pub mod char_stream;
-#[doc(hidden)]
-pub mod dfa_state;
-#[doc(hidden)]
-pub mod interval_set;
-pub mod parser_rule_context;
-mod prediction_context;
-#[doc(hidden)]
-pub mod semantic_context;
-mod token_source;
-pub mod token_stream;
-//pub mod trace_listener;
-#[doc(hidden)]
-pub mod dfa;
-#[doc(hidden)]
-pub mod transition;
-pub mod tree;
-//pub mod file_stream;
-#[doc(hidden)]
 pub mod atn;
-pub mod common_token_stream;
-pub mod error_listener;
-pub mod error_strategy;
-pub mod errors;
-pub mod input_stream;
-pub mod lexer;
-#[doc(hidden)]
-pub mod lexer_action_executor;
-pub mod lexer_atn_simulator;
-pub mod parser;
-pub mod parser_atn_simulator;
-mod prediction_mode;
-pub mod token;
-mod utils;
-//pub mod tokenstream_rewriter_test;
-// mod context_factory;
-pub mod rule_context;
-pub mod vocabulary;
-//#[cfg(test)]
-// tests are either integration tests in "tests" foulder or unit tests in some modules
-
-use std::rc::Rc;
-/// Stable workaround for CoerceUnsized
-// #[doc(hidden)]
-pub trait CoerceFrom<T> {
-    fn coerce_rc(from: Rc<T>) -> Rc<Self>;
-    fn coerce_box(from: Box<T>) -> Box<Self>;
-    fn coerce_ref(from: &T) -> &Self;
-    fn coerce_mut(from: &mut T) -> &mut Self;
-}
-
-#[doc(hidden)]
-#[macro_export]
-macro_rules! coerce_from {
-    ($lt:lifetime : $p:path) => {
-        const _: () = {
-            use std::rc::Rc;
-            impl<$lt, T> $crate::CoerceFrom<T> for dyn $p + $lt
-            where
-                T: $p + $lt,
-            {
-                fn coerce_rc(from: Rc<T>) -> Rc<Self> {
-                    from as _
-                }
-                fn coerce_box(from: Box<T>) -> Box<Self> {
-                    from as _
-                }
-                fn coerce_ref(from: &T) -> &Self {
-                    from as _
-                }
-                fn coerce_mut(from: &mut T) -> &mut Self {
-                    from as _
-                }
-            }
-        };
-    };
-}
-
-/// Stable workaround for CoerceUnsized
-// #[doc(hidden)]
-pub trait CoerceTo<T: ?Sized> {
-    fn coerce_rc_to(self: Rc<Self>) -> Rc<T>;
-    fn coerce_box_to(self: Box<Self>) -> Box<T>;
-    fn coerce_ref_to(&self) -> &T;
-    fn coerce_mut_to(&mut self) -> &mut T;
-}
-
-impl<T: ?Sized, X> CoerceTo<T> for X
-where
-    T: CoerceFrom<X>,
-{
-    fn coerce_rc_to(self: Rc<Self>) -> Rc<T> {
-        T::coerce_rc(self)
-    }
-    fn coerce_box_to(self: Box<Self>) -> Box<T> {
-        T::coerce_box(self)
-    }
-
-    fn coerce_ref_to(&self) -> &T {
-        T::coerce_ref(self)
-    }
-
-    fn coerce_mut_to(&mut self) -> &mut T {
-        T::coerce_mut(self)
-    }
-}
-
-pub mod const_check {
-    pub const fn version(maj: &str, min: &str) -> bool {
-        if maj.is_empty() || min.is_empty() {
-            return false;
-        }
-        byte_eq(VERSION_MAJ.as_bytes(), maj.as_bytes())
-            && byte_eq(VERSION_MIN.as_bytes(), min.as_bytes())
-    }
-
-    const VERSION_MAJ: &str = env!("CARGO_PKG_VERSION_MAJOR");
-    const VERSION_MIN: &str = env!("CARGO_PKG_VERSION_MINOR");
-
-    const fn byte_eq(lhs: &[u8], rhs: &[u8]) -> bool {
-        if lhs.len() != rhs.len() {
-            return false;
-        }
-        let mut i: usize = 0;
-        loop {
-            if i >= lhs.len() {
-                return true;
-            }
-            if lhs[i] != rhs[i] {
-                return false;
-            }
-            i += 1;
-        }
-    }
-}
+pub mod lex;
+pub mod error;
+pub mod utils;
